Move Samples pending for workset queue to Queues page
