+++ run_dir/design/base.html
@@ -83,6 +83,7 @@
           <li class="nav-item dropdown px-1">
             <a class="nav-link dropdown-toggle" id="nav_queues_dropdown" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" href="#">Queues</a>
             <div class="dropdown-menu" aria-labelledby="nav_queues_dropdown" role="menu">
+              <a class="dropdown-item" href="/workset_queues">Pending for Workset queues</a>
               <a class="dropdown-item" href="/pools_qpcr">qPCR Pooling</a>
               <a class="dropdown-item" href="/sequencing_queues">Sequencing queues</a>
             </div>
