+++ run_dir/design/workset_queues.html
@@ -1,0 +1,56 @@
+{% extends "base.html" %}
+
+<!--
+Template file: workset_queues.html
+URL: /workset_queues
+Title: List of workset queues
+Description: Shows a table with extant workset queues.
+-->
+
+{% block stuff %}
+
+<div class="container-fluid" id="worksetQueues_list">
+  <div id="pools_table_div">
+      <h1><span id="page_title">Workset queues</span>
+        <small class="text-muted">All samples that are available to be added to a workset in LIMS.</small>
+      </h1>
+      <button type="button" id="samples_copy_table_btn" class="btn btn-outline-dark"><span class="fa fa-copy"></span> Copy table</button>
+      <button type="button" id="samples_copy_table" class="btn d-none" data-clipboard-target="#samples_table"></button>
+      <label class="btn btn-outline-dark expand-all float-right mb-2"><span class="fa fa-plus-square" aria-hidden="true"></span>Expand All</label>
+      <table class="table table-hover table-striped table-bordered sortable" id="samples_table">
+        <thead id="samples_table_head">
+            <tr class="sticky darkth">
+              <th></th>
+              <th>Project</th>
+              <th>Number of Samples (Total samples in project)</th>
+              <th>Oldest sample (in days)</th>
+              <th>Project Queue Date</th>
+              <th>Has Requeued Samples</th>
+              <th>Lab protocol</th>
+              <th>Latest running note</th>
+            </tr>
+        </thead>
+        <tfoot id="samples_table_footer">
+          <tr class="darkth">
+            <th></th>
+            <th>Project</th>
+            <th>Number of Samples (Total samples in project)</th>
+            <th>Oldest sample (in days)</th>
+            <th>Project Queue Date</th>
+            <th>Has Requeued Samples</th>
+            <th>Lab protocol</th>
+            <th>Latest running note</th>
+          </tr>
+        </tfoot>
+        <tbody class="list" id="samples_table_body">
+        </tbody>
+      </table>
+    </div>
+</div>
+
+<script src="/static/js/clipboard.min.js"></script>
+<script src="/static/js/jquery.dataTables.min.js"></script>
+<script src="/static/js/workset_queues.js?v={{ gs_globals['git_commit'] }}"></script>
+<script src="/static/js/jquery-ui.min.js"></script>
+
+{% end %}
